adding pre-checks based on regex to fail the tests
