+++ .github/workflows/pr-e2e.yml
@@ -92,6 +92,22 @@ jobs:
         run: |
           gh pr checkout ${{ needs.triage.outputs.pr_num }}
 
+
+      - name: Run pre-regex checks
+        continue-on-error: false
+        id: test-regex
+        env:
+          COMMENT_BODY: ${{ github.event.comment.body }}
+        run: |
+          MESSAGE="$COMMENT_BODY"
+          REGEX='/run-e2e (.+)'
+          if [[ "$MESSAGE" =~ $REGEX ]]
+          then
+            export E2E_TEST_REGEX="$(echo ${BASH_REMATCH[1]} | head -1)"
+          fi
+          echo "${{ needs.triage.outputs.pr_num }}"
+          make e2e-regex
+      
       - name: Login to GitHub Container Registry
         uses: docker/login-action@v2
         with:
@@ -155,14 +171,7 @@ jobs:
           GCP_RUN_IDENTITY_TESTS: true
           E2E_IMAGE_TAG: ${{ needs.triage.outputs.image_tag }}
           TEST_CLUSTER_NAME: keda-e2e-cluster-pr
-          COMMENT_BODY: ${{ github.event.comment.body }}
         run: |
-          MESSAGE="$COMMENT_BODY"
-          REGEX='/run-e2e (.+)'
-          if [[ "$MESSAGE" =~ $REGEX ]]
-          then
-            export E2E_TEST_REGEX="$(echo ${BASH_REMATCH[1]} | head -1)"
-          fi
           echo "${{ needs.triage.outputs.pr_num }}"
           make e2e-test
